% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.check_patient_visit.r
\name{ds.check_patient_visit}
\alias{ds.check_patient_visit}
\title{Check patient-visit combinations on a server-side data frame}
\usage{
ds.check_patient_visit(df, datasources = NULL)
}
\arguments{
\item{df}{A character string specifying the name of the server-side data frame
containing at least the columns \code{pat_ID} and \code{Visit_months_from_diagnosis}.}

\item{datasources}{A list of \code{\link[DSI]{DSConnection-class}} objects obtained
after login. If the \code{datasources} argument is not specified, the default set of
connections will be used: see \code{\link[DSI]{datashield.connections_default}}.}
}
\value{
Invisible \code{NULL}. Messages are printed to indicate whether
duplicate patient-visit pairs were found on each server.
}
\description{
Verify all patient-visit combinations in a server-side data frame
and report any duplicates using the \code{check_patient_visitDS} server-side function.
Messages are printed for each server indicating whether duplicates are present.

Server function called: \code{check_patient_visitDS}
}
