% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.remove_columns.r
\name{ds.remove_columns}
\alias{ds.remove_columns}
\title{Remove specified columns from a server-side data frame}
\usage{
ds.remove_columns(df, col_names, newobj = NULL, datasources = NULL)
}
\arguments{
\item{df}{A character string specifying the name of the server-side data frame
from which columns will be removed.}

\item{col_names}{A character vector specifying the names of the columns to remove.
These columns must exist in the data frame, otherwise the server-side function
will throw an error.}

\item{newobj}{A character string specifying the name of the new server-side object
to store the modified data frame. If NULL, a name is generated automatically by
appending \code{"_filtered"} to \code{df}.}

\item{datasources}{a list of \code{\link[DSI]{DSConnection-class}}
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link[DSI]{datashield.connections_default}}.}
}
\value{
Returns the name of the newly created server-side object
containing the modified data frame.
}
\description{
Remove one or more columns from a server-side data frame using
the \code{remove_columnsDS} server-side function. The resulting data frame is
assigned as a new object on each server.

Server function called: \code{remove_columnsDS}
}
